Added header file name extraction on download. Fixed teleport bug. Mortar pickup model points to "ammo/mortar". Bumped protocol version (261 -> 262). Tweaked trail particle effect of several weapon, and readded trail for arrows. Several coop survival changes.

+++ fpsgame/scoreboard.cpp
@@ -153,8 +153,8 @@ namespace game
         g.poplist();
 
 		extern int spawnwait;
-		int spawnms = float(cmode? cmode->respawnmillis(player1): spawnwait-float(lastmillis-player1->lastpain))/1000.f;
-        if(player1->state==CS_DEAD && spawnms>0)
+		float spawnms = float(cmode? cmode->respawnmillis(player1): spawnwait-float(lastmillis-player1->lastpain))/1000.f;
+        if(player1->state==CS_DEAD && spawnms>0.f)
 		{
 			//g.separator();
 			g.titlef("spawn in %.1f", 0xFFFF80, NULL, spawnms);
